requirements.txt
bottle~=0.12.18
uuid~=1.30
python-dotenv~=0.14.0
pony~=0.7.13